pyproject.toml
[tool.poetry]
name = "pii"
version = "0.1.0"
description = ""
authors = ["Dennis Seah <dennis.seah@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
presidio-analyzer = "^2.2.355"
presidio-anonymizer = "^2.2.355"
spacy = "^3.8.2"
lagom = "^2.6.0"
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
faker = "^30.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pyright = "^1.1.386"
pre-commit = "^4.0.1"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "pii_anonymizer/protocols/*",
    "pii_anonymizer/hosting.py"
]